[REFACTOR]: bidPage.html 내부에서 사용하는 변수는 각각의 js 파일 내부에 선언하도록 변경
