+++ pickme/src/main/java/project/pickme/bid/domain/Bid.java
@@ -2,6 +2,7 @@ package project.pickme.bid.domain;
 
 import java.time.LocalDateTime;
 
+import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -11,8 +12,8 @@ import project.pickme.user.domain.User;
 
 @Getter
 @Builder
-@AllArgsConstructor
-@NoArgsConstructor
+@AllArgsConstructor(access = AccessLevel.PROTECTED)
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class Bid {
 	private Long id;
 	private long price;
