+++ pickme/src/main/java/project/pickme/user/repository/CustomsMapper.java
@@ -10,7 +10,10 @@ import project.pickme.user.domain.Customs;
 @Mapper
 public interface CustomsMapper {
 	List<Customs> findAll();
+
 	Optional<Customs> findById(String id);
+
 	void saveAll(List<Customs> customs);
+
 	void deleteAll();
 }
