+++ pickme/src/main/java/project/pickme/user/service/MailService.java
@@ -1,4 +1,4 @@
-package project.pickme.user.controller;
+package project.pickme.user.service;
 
 import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.stereotype.Service;
@@ -31,12 +31,12 @@ public class MailService {
 		message.addRecipients(MimeMessage.RecipientType.TO, toMail);
 		message.setSubject("공매 낙찰 안내");
 		message.setFrom(fromMail);
-		message.setText(setContext(selectedBidDto),"utf-8", "html");
+		message.setText(buildEmailContent(selectedBidDto),"utf-8", "html");
 
 		return message;
 	}
 
-	private String setContext(SelectedBidDto selectedBidDto){
+	private String buildEmailContent(SelectedBidDto selectedBidDto){
 		Context context = new Context();
 		context.setVariable("itemName", selectedBidDto.getItemName());
 		context.setVariable("imageUrl", "https://pickmepickme.s3.ap-northeast-2.amazonaws.com/computer.jpg"); //TODO: selectedBidDto.getImageUrl로 수정하기
