+++ pickme/src/main/resources/templates/charge/charge.html
@@ -119,7 +119,6 @@
   </div>
 </main>
 
-
 <!-- 푸터 -->
 <div th:replace="~{fragments/footer :: footer}">footer content</div>
 </body>
